Added Tray functionality and Tray Icon
